+++ projeto-pomodoro/chronos-pomodoro/src/contexts/TaskContext/TaskContext.tsx
@@ -1,0 +1,16 @@
+import { createContext, type Dispatch } from 'react';
+import { initialTaskState } from './initialTaskState';
+import type { TaskStateModel } from '../../models/TaskStateModel';
+import type { TaskActionModel } from './taskActions';
+
+type TaskContextProps = {
+  state: TaskStateModel;
+  dispatch: Dispatch<TaskActionModel>;
+};
+
+const initialContextValue = {
+  state: initialTaskState,
+  dispatch: () => {},
+};
+
+export const TaskContext = createContext<TaskContextProps>(initialContextValue);
