+++ projeto-pomodoro/chronos-pomodoro/src/contexts/TaskContext/TaskContextProvider.tsx
@@ -1,20 +1,20 @@
-import { useEffect, useState } from 'react';
-import { TaskContext } from '.';
+import { useEffect, useReducer } from 'react';
 import { initialTaskState } from './initialTaskState';
+import { taskReducer } from './taskReducer';
+import { TaskContext } from './TaskContext';
 
 type TaskContextProviderProps = {
   children: React.ReactNode;
 };
 
 export function TaskContextProvider({ children }: TaskContextProviderProps) {
-  const [state, setState] = useState(initialTaskState);
-
+  const [state, dispatch] = useReducer(taskReducer, initialTaskState);
   useEffect(() => {
     console.log(state);
   }, [state]);
 
   return (
-    <TaskContext.Provider value={{ state, setState }}>
+    <TaskContext.Provider value={{ state, dispatch }}>
       {children}
     </TaskContext.Provider>
   );
