+++ projeto-pomodoro/chronos-pomodoro/src/contexts/TaskContext/index.tsx
@@ -1,48 +1,8 @@
-import { createContext, useContext, useEffect, useState } from 'react';
-import type { TaskStateModel } from '../../models/TaskStateModel';
+import { useContext } from 'react';
+import { TaskContext } from './TaskContext';
 
-const initialState: TaskStateModel = {
-  tasks: [],
-  secondsRemaining: 0,
-  formattedSecondsRemaining: '00:00',
-  activeTask: null,
-  currentCycle: 0,
-  config: {
-    workTime: 25,
-    shortBreakTime: 5,
-    longBreakTime: 15,
-  },
-};
-
-type TaskContextProps = {
-  state: TaskStateModel;
-  setState: React.Dispatch<React.SetStateAction<TaskStateModel>>;
-};
-
-const initialContextValue = {
-  state: initialState,
-  setState: () => {},
-};
-
-export const TaskContext = createContext<TaskContextProps>(initialContextValue);
-
-type TaskContextProviderProps = {
-  children: React.ReactNode;
-};
-
-export function TaskContextProvider({ children }: TaskContextProviderProps) {
-  const [state, setState] = useState(initialState);
-
-  useEffect(() => {
-    console.log(state);
-  }, [state]);
-
-  return (
-    <TaskContext.Provider value={{ state, setState }}>
-      {children}
-    </TaskContext.Provider>
-  );
-}
+export { TaskContext } from './TaskContext';
+export { TaskContextProvider } from './TaskContextProvider';
 
 export function useTaskContext() {
   return useContext(TaskContext);
