+++ projeto-pomodoro/chronos-pomodoro/src/contexts/TaskContext/taskActions.ts
@@ -14,20 +14,21 @@ export const TaskActionTypes = {
   RESET_STATE: 'RESET_STATE',
 } as const;
 
-export type TaskActionsWithPayload =
+export type TaskActionTypes =
+  (typeof TaskActionTypes)[keyof typeof TaskActionTypes];
+
+export type TaskActionsWithPayload = {
+  type: typeof TaskActionTypes.START_TASK;
+  payload: TaskModel;
+};
+
+export type TaskActionsWithoutPayload =
   | {
-      // 2. Adicionamos 'typeof' antes do valor
-      type: typeof TaskActionTypes.START_TASK;
-      payload: TaskModel;
+      type: typeof TaskActionTypes.RESET_STATE;
     }
   | {
       type: typeof TaskActionTypes.INTERRUPT_TASK;
-      payload: TaskModel;
     };
-
-export type TaskActionsWithoutPayload = {
-  type: typeof TaskActionTypes.RESET_STATE;
-};
 
 export type TaskActionModel =
   | TaskActionsWithPayload
